move RenamingListAttribute to Manager
