+++ src/es6/canvas_widget/viewpoint/ViewRelationshipNode.js
@@ -2,7 +2,7 @@ import "https://unpkg.com/jquery@3.6.0/dist/jquery.js";
 import "https://cdnjs.cloudflare.com/ajax/libs/jqueryui/1.13.2/jquery-ui.min.js";
 import _ from "lodash-es";
 import { SingleSelectionAttribute } from "../Manager";
-import RenamingListAttribute from "./RenamingListAttribute";
+import { RenamingListAttribute } from "../Manager";
 import ConditionListAttribute from "./ConditionListAttribute";
 import ViewTypesUtil from "./ViewTypesUtil";
 import LogicalOperator from "./LogicalOperator";
